Desktop/deepfake-detector/app.py
import os
import cv2
import numpy as np
import torch
import base64
from io import BytesIO
from PIL import Image, ImageDraw
from flask import Flask, request, jsonify, render_template, send_from_directory
from flask_cors import CORS
from transformers import AutoModelForImageClassification, AutoFeatureExtractor
from werkzeug.utils import secure_filename

# Initialize Flask app
app = Flask(__name__)
CORS(app)  # Enable CORS for all routes

# Configure upload folder
UPLOAD_FOLDER = 'uploads'
ALLOWED_EXTENSIONS = {'mp4', 'avi', 'mov', 'jpg', 'jpeg', 'png','webp'}
app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
app.config['MAX_CONTENT_LENGTH'] = 50 * 1024 * 1024  # 50MB max upload

# Create uploads directory if it doesn't exist
os.makedirs(UPLOAD_FOLDER, exist_ok=True)

# Load DeepFake Detection Model
print("Loading model...")
model_name = "dima806/deepfake_vs_real_image_detection"
model = AutoModelForImageClassification.from_pretrained(model_name)
extractor = AutoFeatureExtractor.from_pretrained(model_name)
print("Model loaded successfully.")

def allowed_file(filename):
    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS

# Function to extract frames from video
def extract_frames(video_path, frame_rate=1):
    # Your existing extract_frames function
    cap = cv2.VideoCapture(video_path)
    frames = []
    frame_indices = []
    
    if not cap.isOpened():
        return [], []
    
    fps = int(cap.get(cv2.CAP_PROP_FPS))
    total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
    duration = total_frames / fps if fps > 0 else 0
    
    # Extract frames at the specified rate
    interval = max(1, int(fps / frame_rate))
    count = 0
    
    while cap.isOpened():
        ret, frame = cap.read()
        if not ret:
            break
        
        if count % interval == 0:
            rgb_frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
            frames.append(rgb_frame)
            frame_indices.append(count)
        
        count += 1
    
    cap.release()
    return frames, frame_indices, duration, fps

# Function to classify each frame
def classify_frames(frames):
    # Your existing classify_frames function
    deepfake_scores = []
    
    for frame in frames:
        # Convert numpy array to PIL Image
        pil_frame = Image.fromarray(frame)
        
        # Prepare input for the model
        inputs = extractor(images=pil_frame, return_tensors="pt")
        
        # Get model prediction
        with torch.no_grad():
            outputs = model(**inputs)
            scores = torch.nn.functional.softmax(outputs.logits, dim=-1)
            deepfake_score = scores[0][1].item()  # Probability of being fake
            deepfake_scores.append(deepfake_score)
    
    return deepfake_scores

# Function to generate heatmap
def generate_heatmap(frame, score):
    # Your existing generate_heatmap function
    img = Image.fromarray(frame)
    overlay = Image.new('RGBA', img.size, (0, 0, 0, 0))
    draw = ImageDraw.Draw(overlay)
    
    # Create a transparent red overlay with opacity based on score
    red_alpha = int(155 * score)  # Higher score = more opaque
    draw.rectangle([(0, 0), img.size], fill=(255, 0, 0, red_alpha))
    
    # Blend the original image with the overlay
    result = Image.alpha_composite(img.convert('RGBA'), overlay)
    return result

# Function to encode image to base64
def encode_image_to_base64(image):
    # Your existing encode_image_to_base64 function
    buffered = BytesIO()
    image.save(buffered, format="PNG")
    return base64.b64encode(buffered.getvalue()).decode('utf-8')

# Function to analyze an image
def analyze_image(image_path):
    # Your existing analyze_image function
    # Read and convert image
    img = cv2.imread(image_path)
    if img is None:
        return {"error": "Could not read image file"}
    
    rgb_img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
    
    # Classify the image
    pil_img = Image.fromarray(rgb_img)
    inputs = extractor(images=pil_img, return_tensors="pt")
    
    with torch.no_grad():
        outputs = model(**inputs)
        scores = torch.nn.functional.softmax(outputs.logits, dim=-1)
        fake_score = scores[0][1].item()
    
    # Generate heatmap
    heatmap = generate_heatmap(rgb_img, fake_score)
    heatmap_base64 = encode_image_to_base64(heatmap)
    
    # Prepare detailed analysis
    analysis = {
        "confidence": fake_score * 100,  # Convert to percentage
        "classification": "Potential Deepfake" if fake_score > 0.5 else "Likely Authentic",
        "original_image": encode_image_to_base64(pil_img),
        "heatmap_image": heatmap_base64,
        "details": {
            "file_name": os.path.basename(image_path),
            "analysis_duration": 0.5,  # Placeholder
            "ai_confidence": fake_score * 100
        },
        "analysis_sections": {
            "facial_analysis": {
                "status": "warning" if fake_score > 0.4 else "pass",
                "description": "Potential facial inconsistencies detected." if fake_score > 0.4 else "No significant facial anomalies detected."
            },
            "frequency_analysis": {
                "status": "fail" if fake_score > 0.6 else "warning" if fake_score > 0.4 else "pass",
                "description": "GAN artifacts detected in frequency domain." if fake_score > 0.6 else 
                              "Some unusual patterns in frequency domain." if fake_score > 0.4 else 
                              "No significant frequency domain anomalies."
            }
        }
    }
    
    return analysis

# Function to analyze the video
def analyze_video(video_path):
    # Your existing analyze_video function
    # Extract frames from the video
    frames, frame_indices, duration, fps = extract_frames(video_path, frame_rate=1)
    
    if not frames:
        return {"error": "Failed to extract frames from the video"}
    
    # Analyze each frame
    scores = classify_frames(frames)
    
    # Generate heatmap for the most suspicious frame
    max_score_index = np.argmax(scores)
    max_score = scores[max_score_index]
    heatmap = generate_heatmap(frames[max_score_index], max_score)
    
    # Convert original frame and heatmap to base64 for frontend display
    original_frame = Image.fromarray(frames[max_score_index])
    original_base64 = encode_image_to_base64(original_frame)
    heatmap_base64 = encode_image_to_base64(heatmap)
    
    # Determine overall classification
    avg_score = np.mean(scores)
    max_score = np.max(scores)
    min_score = np.min(scores)
    
    # Calculate time-based metrics
    time_above_threshold = sum(1 for s in scores if s > 0.5) / len(scores)
    
    # Prepare detailed analysis
    analysis = {
        "confidence": avg_score * 100,  # Convert to percentage
        "classification": "Potential Deepfake" if avg_score > 0.5 else "Likely Authentic",
        "original_frame": original_base64,
        "heatmap_frame": heatmap_base64,
        "details": {
            "file_name": os.path.basename(video_path),
            "analysis_duration": len(frames) * 0.3,  # Approximate time taken
            "media_length": f"{duration:.1f} seconds",
            "ai_confidence": f"{avg_score * 100:.1f}%"
        },
        "analysis_sections": {
            "facial_analysis": {
                "status": "fail" if max_score > 0.7 else "warning" if max_score > 0.5 else "pass",
                "description": "Unnatural facial movements detected." if max_score > 0.7 else 
                              "Some unusual facial patterns detected." if max_score > 0.5 else 
                              "No significant facial anomalies detected."
            },
            "frequency_analysis": {
                "status": "fail" if avg_score > 0.6 else "warning" if avg_score > 0.4 else "pass",
                "description": "GAN artifacts detected in frequency domain." if avg_score > 0.6 else 
                              "Some unusual patterns in frequency domain." if avg_score > 0.4 else 
                              "No significant frequency domain anomalies."
            },
            "audio_visual_sync": {
                "status": "warning" if time_above_threshold > 0.5 else "pass",
                "description": "Potential audio-visual sync issues detected." if time_above_threshold > 0.5 else 
                              "Audio synchronization appears normal."
            },
            "lighting_consistency": {
                "status": "fail" if max_score - min_score > 0.3 else "warning" if max_score - min_score > 0.15 else "pass",
                "description": "Inconsistent lighting detected across frames." if max_score - min_score > 0.3 else 
                              "Slight lighting inconsistencies detected." if max_score - min_score > 0.15 else 
                              "Lighting appears consistent throughout the video."
            }
        },
        "frame_analysis": [
            {
                "frame_index": idx,
                "time": f"{idx/fps:.2f}s",
                "score": score * 100
            } for idx, score in zip(frame_indices, scores)
        ]
    }
    
    return analysis

# Routes for serving HTML pages
@app.route('/')
def index():
    return render_template('index.html')

@app.route('/results.html')
@app.route('/results')
def results():
    return render_template('results.html')

# API route for analyzing media
@app.route('/api/analyze', methods=['POST'])
def analyze_media():
    # Check if the post request has the file part
    if 'file' not in request.files:
        return jsonify({"error": "No file part"}), 400
    
    file = request.files['file']
    
    # If the user does not select a file, the browser submits an empty file without a filename
    if file.filename == '':
        return jsonify({"error": "No selected file"}), 400
    
    if file and allowed_file(file.filename):
        filename = secure_filename(file.filename)
        file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
        file.save(file_path)
        
        # Determine if it's an image or video and process accordingly
        if filename.lower().endswith(('.jpg', '.jpeg', '.png')):
            result = analyze_image(file_path)
        else:
            result = analyze_video(file_path)
        
        # Clean up the file after processing
        os.remove(file_path)
        
        return jsonify(result)
    
    return jsonify({"error": "File type not allowed"}), 400

# Route for serving static files (JS, CSS, etc.)
@app.route('/static/<path:path>')
def serve_static(path):
    return send_from_directory('static', path)

if __name__ == '__main__':
    app.run(debug=True, host='0.0.0.0', port=5000)